int
id yo ibt

+++ carRentForCompany2/app/src/main/java/com/example/moish/carrentforcompany/model/entities/Client.java
@@ -12,7 +12,7 @@ public class Client {
     int id;
     String phoneNumber;
     String email;
-    long creditCardNumber;
+    int creditCardNumber;
 
     public Client() {
     }
